refactor: Cambiar la base de las rutas de 'my-account' a 'purchases'
- Se ha renombrado el directorio de las páginas del frontend de 'my-account' a 'purchases'.
- Se han actualizado los enlaces de navegación para que apunten a las nuevas rutas.
- Se han actualizado las rutas de la API del backend para que usen la base 'purchases'.
- Se han actualizado las llamadas a la API en los componentes de Vue para que coincidan con las nuevas rutas del backend.
- Se ha renombrado el archivo de rutas del backend a 'purchases-routes.php' y se ha actualizado su inclusión.
