+++ wp-content/plugins/motorlan-api-vue/app/src/pages/apps/purchases/favorites.vue
@@ -1,13 +1,14 @@
 <script setup lang="ts">
 import { useRouter } from 'vue-router'
-import type { ImagenDestacada } from '../../../../../interfaces/publicacion'
+import type { ImagenDestacada, Publicacion } from '../../../../../interfaces/publicacion'
 
 const router = useRouter()
 
 const headers = [
-  { title: 'Publicacion', key: 'motor' },
-  { title: 'Fecha de Compra', key: 'fecha_compra' },
-  { title: 'Estado', key: 'estado' },
+  { title: 'Publicacion', key: 'title' },
+  { title: 'Referencia', key: 'referencia' },
+  { title: 'Precio', key: 'precio' },
+  { title: 'Estado', key: 'status' },
   { title: 'Acciones', key: 'actions', sortable: false },
 ]
 
@@ -26,18 +27,7 @@ const updateOptions = (options: any) => {
   orderBy.value = options.sortBy[0]?.order
 }
 
-const resolveStatus = (status: string) => {
-  if (status === 'pendiente')
-    return { text: 'Pendiente', color: 'warning' }
-  if (status === 'completado')
-    return { text: 'Completado', color: 'success' }
-  if (status === 'cancelado')
-    return { text: 'Cancelado', color: 'error' }
-
-  return { text: 'Desconocido', color: 'info' }
-}
-
-const { data: purchasesData } = await useApi<any>(createUrl('/wp-json/motorlan/v1/my-account/purchases', {
+const { data: favoritesData, execute: fetchFavorites } = await useApi<any>(createUrl('/wp-json/motorlan/v1/purchases/favorites', {
   query: {
     page,
     per_page: itemsPerPage,
@@ -47,8 +37,31 @@ const { data: purchasesData } = await useApi<any>(createUrl('/wp-json/motorlan/v
   },
 }))
 
-const purchases = computed(() => purchasesData.value?.data || [])
-const totalPurchases = computed(() => purchasesData.value?.pagination.total || 0)
+const favorites = computed((): Publicacion[] => favoritesData.value?.data || [])
+const totalFavorites = computed(() => favoritesData.value?.pagination.total || 0)
+
+const resolveStatus = (status: string) => {
+  if (status === 'publish')
+    return { text: 'Publicado', color: 'success' }
+  if (status === 'draft')
+    return { text: 'Borrador', color: 'secondary' }
+  if (status === 'paused')
+    return { text: 'Pausado', color: 'warning' }
+  if (status === 'sold')
+    return { text: 'Vendido', color: 'error' }
+
+  return { text: 'Desconocido', color: 'info' }
+}
+
+const removeFavorite = async (motorId: number) => {
+  try {
+    await $api(`/wp-json/motorlan/v1/purchases/favorites/${motorId}`, { method: 'DELETE' })
+    fetchFavorites()
+  }
+  catch (error) {
+    console.error('Error removing favorite:', error)
+  }
+}
 
 const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail'): string => {
   let imageObj: ImagenDestacada | null = null
@@ -70,8 +83,8 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
 
 <template>
   <VCard
-    id="purchase-list"
-    title="Mis Compras"
+    id="favorite-list"
+    title="Mis Favoritos"
   >
     <VCardText>
       <div class="d-flex flex-wrap gap-4">
@@ -79,7 +92,7 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
           <!-- 👉 Search  -->
           <AppTextField
             v-model="searchQuery"
-            placeholder="Buscar Compra"
+            placeholder="Buscar Favorito"
             style="inline-size: 200px;"
             class="me-3"
           />
@@ -102,43 +115,45 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
       v-model:items-per-page="itemsPerPage"
       v-model:page="page"
       :headers="headers"
-      :items="purchases"
-      :items-length="totalPurchases"
+      :items="favorites"
+      :items-length="totalFavorites"
       class="text-no-wrap"
       @update:options="updateOptions"
     >
       <!-- publicacion -->
-      <template #item.motor="{ item }">
-        <div
-          v-if="item.motor"
-          class="d-flex align-center gap-x-4"
-        >
+      <template #item.title="{ item }">
+        <div class="d-flex align-center gap-x-4">
           <VAvatar
-            v-if="item.motor.imagen_destacada"
+            v-if="item.imagen_destacada"
             size="38"
             variant="tonal"
             rounded
-            :image="getImageBySize(item.motor.imagen_destacada, 'thumbnail')"
+            :image="getImageBySize(item.imagen_destacada, 'thumbnail')"
           />
           <div class="d-flex flex-column">
             <span
               class="text-body-1 font-weight-medium text-high-emphasis cursor-pointer"
-              @click="router.push(`/apps/motors/motor/edit/${item.motor.uuid}`)"
-            >{{ item.motor.title }}</span>
-            <span class="text-body-2">{{ item.motor.acf.marca?.name }}</span>
+              @click="router.push(`/apps/motors/motor/edit/${item.uuid}`)"
+            >{{ item.title }}</span>
+            <span class="text-body-2">{{ item.acf.marca?.name }}</span>
           </div>
         </div>
       </template>
 
-      <!-- fecha_compra -->
-      <template #item.fecha_compra="{ item }">
-        <span class="text-body-1 text-high-emphasis">{{ item.fecha_compra }}</span>
+      <!-- referencia -->
+      <template #item.referencia="{ item }">
+        <span class="text-body-1 text-high-emphasis">{{ item.acf.tipo_o_referencia }}</span>
       </template>
 
-      <!-- estado -->
-      <template #item.estado="{ item }">
+      <!-- precio -->
+      <template #item.precio="{ item }">
+        <span class="text-body-1 text-high-emphasis">{{ item.acf.precio_de_venta }}</span>
+      </template>
+
+      <!-- status -->
+      <template #item.status="{ item }">
         <VChip
-          v-bind="resolveStatus(item.estado)"
+          v-bind="resolveStatus(item.status)"
           density="default"
           label
           size="small"
@@ -147,8 +162,11 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
 
       <!-- Actions -->
       <template #item.actions="{ item }">
-        <IconBtn @click="router.push(`/apps/purchases/view/${item.uuid}`)">
+        <IconBtn @click="router.push(`/apps/motors/motor/edit/${item.uuid}`)">
           <VIcon icon="tabler-eye" />
+        </IconBtn>
+        <IconBtn @click="removeFavorite(item.id)">
+          <VIcon icon="tabler-trash" />
         </IconBtn>
       </template>
 
@@ -157,7 +175,7 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
         <TablePagination
           v-model:page="page"
           :items-per-page="itemsPerPage"
-          :total-items="totalPurchases"
+          :total-items="totalFavorites"
         />
       </template>
     </VDataTableServer>
