+++ wp-content/plugins/motorlan-api-vue/app/src/pages/apps/purchases/index.vue
