+++ wp-content/plugins/motorlan-api-vue/app/src/pages/apps/purchases/questions.vue
@@ -1,14 +1,14 @@
 <script setup lang="ts">
 import { useRouter } from 'vue-router'
-import type { ImagenDestacada, Publicacion } from '../../../../../interfaces/publicacion'
+import type { ImagenDestacada } from '../../../../../interfaces/publicacion'
 
 const router = useRouter()
 
 const headers = [
-  { title: 'Publicacion', key: 'title' },
-  { title: 'Referencia', key: 'referencia' },
-  { title: 'Precio', key: 'precio' },
-  { title: 'Estado', key: 'status' },
+  { title: 'Publicacion', key: 'motor' },
+  { title: 'Pregunta', key: 'pregunta' },
+  { title: 'Respuesta', key: 'respuesta' },
+  { title: 'Estado', key: 'estado' },
   { title: 'Acciones', key: 'actions', sortable: false },
 ]
 
@@ -27,7 +27,7 @@ const updateOptions = (options: any) => {
   orderBy.value = options.sortBy[0]?.order
 }
 
-const { data: favoritesData, execute: fetchFavorites } = await useApi<any>(createUrl('/wp-json/motorlan/v1/my-account/favorites', {
+const { data: questionsData } = await useApi<any>(createUrl('/wp-json/motorlan/v1/purchases/questions', {
   query: {
     page,
     per_page: itemsPerPage,
@@ -37,30 +37,14 @@ const { data: favoritesData, execute: fetchFavorites } = await useApi<any>(creat
   },
 }))
 
-const favorites = computed((): Publicacion[] => favoritesData.value?.data || [])
-const totalFavorites = computed(() => favoritesData.value?.pagination.total || 0)
+const questions = computed(() => questionsData.value?.data || [])
+const totalQuestions = computed(() => questionsData.value?.pagination.total || 0)
 
-const resolveStatus = (status: string) => {
-  if (status === 'publish')
-    return { text: 'Publicado', color: 'success' }
-  if (status === 'draft')
-    return { text: 'Borrador', color: 'secondary' }
-  if (status === 'paused')
-    return { text: 'Pausado', color: 'warning' }
-  if (status === 'sold')
-    return { text: 'Vendido', color: 'error' }
+const resolveStatus = (respuesta: string | null) => {
+  if (respuesta)
+    return { text: 'Respondida', color: 'success' }
 
-  return { text: 'Desconocido', color: 'info' }
-}
-
-const removeFavorite = async (motorId: number) => {
-  try {
-    await $api(`/wp-json/motorlan/v1/my-account/favorites/${motorId}`, { method: 'DELETE' })
-    fetchFavorites()
-  }
-  catch (error) {
-    console.error('Error removing favorite:', error)
-  }
+  return { text: 'Pendiente', color: 'warning' }
 }
 
 const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail'): string => {
@@ -83,8 +67,8 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
 
 <template>
   <VCard
-    id="favorite-list"
-    title="Mis Favoritos"
+    id="question-list"
+    title="Mis Preguntas"
   >
     <VCardText>
       <div class="d-flex flex-wrap gap-4">
@@ -92,7 +76,7 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
           <!-- 👉 Search  -->
           <AppTextField
             v-model="searchQuery"
-            placeholder="Buscar Favorito"
+            placeholder="Buscar Pregunta"
             style="inline-size: 200px;"
             class="me-3"
           />
@@ -115,45 +99,48 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
       v-model:items-per-page="itemsPerPage"
       v-model:page="page"
       :headers="headers"
-      :items="favorites"
-      :items-length="totalFavorites"
+      :items="questions"
+      :items-length="totalQuestions"
       class="text-no-wrap"
       @update:options="updateOptions"
     >
       <!-- publicacion -->
-      <template #item.title="{ item }">
-        <div class="d-flex align-center gap-x-4">
+      <template #item.motor="{ item }">
+        <div
+          v-if="item.motor"
+          class="d-flex align-center gap-x-4"
+        >
           <VAvatar
-            v-if="item.imagen_destacada"
+            v-if="item.motor.imagen_destacada"
             size="38"
             variant="tonal"
             rounded
-            :image="getImageBySize(item.imagen_destacada, 'thumbnail')"
+            :image="getImageBySize(item.motor.imagen_destacada, 'thumbnail')"
           />
           <div class="d-flex flex-column">
             <span
               class="text-body-1 font-weight-medium text-high-emphasis cursor-pointer"
-              @click="router.push(`/apps/motors/motor/edit/${item.uuid}`)"
-            >{{ item.title }}</span>
-            <span class="text-body-2">{{ item.acf.marca?.name }}</span>
+              @click="router.push(`/apps/motors/motor/edit/${item.motor.uuid}`)"
+            >{{ item.motor.title }}</span>
+            <span class="text-body-2">{{ item.motor.acf.marca?.name }}</span>
           </div>
         </div>
       </template>
 
-      <!-- referencia -->
-      <template #item.referencia="{ item }">
-        <span class="text-body-1 text-high-emphasis">{{ item.acf.tipo_o_referencia }}</span>
+      <!-- pregunta -->
+      <template #item.pregunta="{ item }">
+        <span class="text-body-1 text-high-emphasis">{{ item.pregunta }}</span>
       </template>
 
-      <!-- precio -->
-      <template #item.precio="{ item }">
-        <span class="text-body-1 text-high-emphasis">{{ item.acf.precio_de_venta }}</span>
+      <!-- respuesta -->
+      <template #item.respuesta="{ item }">
+        <span class="text-body-1 text-high-emphasis">{{ item.respuesta || 'Sin respuesta' }}</span>
       </template>
 
-      <!-- status -->
-      <template #item.status="{ item }">
+      <!-- estado -->
+      <template #item.estado="{ item }">
         <VChip
-          v-bind="resolveStatus(item.status)"
+          v-bind="resolveStatus(item.respuesta)"
           density="default"
           label
           size="small"
@@ -162,11 +149,11 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
 
       <!-- Actions -->
       <template #item.actions="{ item }">
-        <IconBtn @click="router.push(`/apps/motors/motor/edit/${item.uuid}`)">
+        <IconBtn
+          v-if="item.motor"
+          @click="router.push(`/apps/motors/motor/edit/${item.motor.uuid}`)"
+        >
           <VIcon icon="tabler-eye" />
-        </IconBtn>
-        <IconBtn @click="removeFavorite(item.id)">
-          <VIcon icon="tabler-trash" />
         </IconBtn>
       </template>
 
@@ -175,7 +162,7 @@ const getImageBySize = (image: ImagenDestacada | null | any[], size = 'thumbnail
         <TablePagination
           v-model:page="page"
           :items-per-page="itemsPerPage"
-          :total-items="totalFavorites"
+          :total-items="totalQuestions"
         />
       </template>
     </VDataTableServer>
